Conflicts:
	Source/DataSources/ConeVisualizer.js
	Source/DataSources/CzmlDataSource.js
	Source/DataSources/EllipseGeometryUpdater.js
	Source/DataSources/EllipsoidGeometryUpdater.js
	Source/DataSources/Entity.js
	Source/DataSources/PathVisualizer.js
	Source/DataSources/PolygonGeometryUpdater.js
	Source/DataSources/PolylineGeometryUpdater.js
	Source/DataSources/RectangleGeometryUpdater.js
	Source/DataSources/StaticGeometryColorBatch.js
	Source/DataSources/StaticGeometryPerMaterialBatch.js
	Source/DataSources/StaticOutlineGeometryBatch.js
	Source/DynamicScene/DynamicBillboardVisualizer.js
	Source/DynamicScene/DynamicLabelVisualizer.js
	Source/DynamicScene/DynamicModelVisualizer.js
	Source/DynamicScene/DynamicPointVisualizer.js
	Source/DynamicScene/DynamicPyramidVisualizer.js
	Source/DynamicScene/DynamicVectorVisualizer.js
	Source/Widgets/Viewer/viewerEntityMixin.js
	Specs/DynamicScene/DynamicObjectSpec.js
	Specs/Widgets/Viewer/viewerDynamicObjectMixinSpec.js

+++ Source/Widgets/Viewer/Viewer.js
@@ -7,8 +7,8 @@ define([
         '../../Core/destroyObject',
         '../../Core/DeveloperError',
         '../../Core/EventHelper',
-        '../../DynamicScene/DataSourceCollection',
-        '../../DynamicScene/DataSourceDisplay',
+        '../../DataSources/DataSourceCollection',
+        '../../DataSources/DataSourceDisplay',
         '../../ThirdParty/knockout',
         '../Animation/Animation',
         '../Animation/AnimationViewModel',
@@ -86,6 +86,7 @@ define([
      * @param {Boolean} [options.timeline=true] If set to false, the Timeline widget will not be created.
      * @param {Boolean} [options.navigationHelpButton=true] If set to the false, the navigation help button will not be created.
      * @param {Boolean} [options.navigationInstructionsInitiallyVisible=true] True if the navigation instructions should initially be visible, or false if the should not be shown until the user explicitly clicks the button.
+     * @param {Boolean} [options.scene3DOnly=false] When <code>true</code>, each geometry instance will only be rendered in 3D to save GPU memory.
      * @param {ProviderViewModel} [options.selectedImageryProviderViewModel] The view model for the current base imagery layer, if not supplied the first available base layer is used.  This value is only valid if options.baseLayerPicker is set to true.
      * @param {ProviderViewModel[]} [options.imageryProviderViewModels=createDefaultImageryProviderViewModels()] The array of ProviderViewModels to be selectable from the BaseLayerPicker.  This value is only valid if options.baseLayerPicker is set to true.
      * @param {ProviderViewModel} [options.selectedTerrainProviderViewModel] The view model for the current base terrain layer, if not supplied the first available base layer is used.  This value is only valid if options.baseLayerPicker is set to true.
@@ -102,7 +103,8 @@ define([
      * @param {SceneMode} [options.sceneMode=SceneMode.SCENE3D] The initial scene mode.
      * @param {MapProjection} [options.mapProjection=new GeographicProjection()] The map projection to use in 2D and Columbus View modes.
      * @param {Element|String} [options.creditContainer] The DOM element or ID that will contain the {@link CreditDisplay}.  If not specified, the credits are added to the bottom of the widget itself.
-     * @param {DataSourceCollection} [options.dataSources=new DataSourceCollection()] The collection of data sources visualized by the widget.
+     * @param {DataSourceCollection} [options.dataSources=new DataSourceCollection()] The collection of data sources visualized by the widget.  If this parameter is provided,
+                                     the instance is assumed to be owned by the caller and will not be destroyed when the viewer is destroyed.
      *
      * @exception {DeveloperError} Element with id "container" does not exist in the document.
      * @exception {DeveloperError} options.imageryProvider is not available when using the BaseLayerPicker widget, specify options.selectedImageryProviderViewModel instead.
@@ -118,7 +120,7 @@ define([
      * @see SceneModePicker
      * @see Timeline
      * @see viewerDragDropMixin
-     * @see viewerDynamicObjectMixin
+     * @see viewerEntityMixin
      *
      * @example
      * //Initialize the viewer widget with several custom options and mixins.
@@ -156,7 +158,7 @@ define([
      * viewer.extend(Cesium.viewerDragDropMixin);
      *
      * //Allow users to zoom and follow objects loaded from CZML by clicking on it.
-     * viewer.extend(Cesium.viewerDynamicObjectMixin);
+     * viewer.extend(Cesium.viewerEntityMixin);
      *
      * //Show a pop-up alert if we encounter an error when processing a dropped file
      * viewer.dropError.addEventListener(function(dropHandler, name, error) {
@@ -217,6 +219,8 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
 
         viewerContainer.appendChild(bottomContainer);
 
+        var scene3DOnly = defaultValue(options.scene3DOnly, false);
+
         // Cesium widget
         var cesiumWidget = new CesiumWidget(cesiumWidgetContainer, {
             terrainProvider : options.terrainProvider,
@@ -228,15 +232,21 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
             useDefaultRenderLoop : options.useDefaultRenderLoop,
             targetFrameRate : options.targetFrameRate,
             showRenderLoopErrors : options.showRenderLoopErrors,
-            creditContainer : defined(options.creditContainer) ? options.creditContainer : bottomContainer
+            creditContainer : defined(options.creditContainer) ? options.creditContainer : bottomContainer,
+            scene3DOnly : scene3DOnly
         });
 
         var dataSourceCollection = options.dataSources;
+        var destroyDataSourceCollection = false;
         if (!defined(dataSourceCollection)) {
             dataSourceCollection = new DataSourceCollection();
+            destroyDataSourceCollection = true;
         }
 
-        var dataSourceDisplay = new DataSourceDisplay(cesiumWidget.scene, dataSourceCollection);
+        var dataSourceDisplay = new DataSourceDisplay({
+            scene : cesiumWidget.scene,
+            dataSourceCollection : dataSourceCollection
+        });
 
         var clock = cesiumWidget.clock;
         var clockViewModel = new ClockViewModel(clock);
@@ -281,15 +291,14 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
             toolbar.appendChild(geocoderContainer);
             geocoder = new Geocoder({
                 container : geocoderContainer,
-                scene : cesiumWidget.scene,
-                ellipsoid : cesiumWidget.scene.globe.ellipsoid
+                scene : cesiumWidget.scene
             });
         }
 
         // HomeButton
         var homeButton;
         if (!defined(options.homeButton) || options.homeButton !== false) {
-            homeButton = new HomeButton(toolbar, cesiumWidget.scene, cesiumWidget.scene.globe.ellipsoid);
+            homeButton = new HomeButton(toolbar, cesiumWidget.scene);
             if (defined(geocoder)) {
                 eventHelper.add(homeButton.viewModel.command.afterExecute, function() {
                     var viewModel = geocoder.viewModel;
@@ -302,8 +311,14 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
         }
 
         // SceneModePicker
+        // By default, we silently disable the scene mode picker if scene3DOnly is true,
+        // but if sceneModePicker is explicitly set to true, throw an error.
+        if ((options.sceneModePicker === true) && scene3DOnly) {
+            throw new DeveloperError('options.sceneModePicker is not available when options.scene3DOnly is set to true.');
+        }
+
         var sceneModePicker;
-        if (!defined(options.sceneModePicker) || options.sceneModePicker !== false) {
+        if (!scene3DOnly && (!defined(options.sceneModePicker) || options.sceneModePicker !== false)) {
             sceneModePicker = new SceneModePicker(toolbar, cesiumWidget.scene);
         }
 
@@ -329,9 +344,18 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
         // Navigation Help Button
         var navigationHelpButton;
         if (!defined(options.navigationHelpButton) || options.navigationHelpButton !== false) {
+            var showNavHelp = true;
+            if (defined(window.localStorage)) {
+                var  hasSeenNavHelp = window.localStorage.getItem('cesium-hasSeenNavHelp');
+                if (defined(hasSeenNavHelp) && Boolean(hasSeenNavHelp)) {
+                    showNavHelp = false;
+                } else {
+                    window.localStorage.setItem('cesium-hasSeenNavHelp', 'true');
+                }
+            }
             navigationHelpButton = new NavigationHelpButton({
                 container : toolbar,
-                instructionsInitiallyVisible : defaultValue(options.navigationInstructionsInitiallyVisible, true)
+                instructionsInitiallyVisible : defaultValue(options.navigationInstructionsInitiallyVisible, showNavHelp)
             });
         }
 
@@ -419,14 +443,14 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
             if (automaticallyTrackDataSourceClocks) {
                 that.clockTrackedDataSource = dataSource;
             }
-            var id = dataSource.dynamicObjects.id;
+            var id = dataSource.entities.id;
             var removalFunc = eventHelper.add(dataSource.changedEvent, onDataSourceChanged);
             that._dataSourceChangedListeners[id] = removalFunc;
         };
 
         var onDataSourceRemoved = function(dataSourceCollection, dataSource) {
             var resetClock = (that.clockTrackedDataSource === dataSource);
-            var id = dataSource.dynamicObjects.id;
+            var id = dataSource.entities.id;
             that._dataSourceChangedListeners[id]();
             that._dataSourceChangedListeners[id] = undefined;
             if (resetClock) {
@@ -461,6 +485,7 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
         this._selectionIndicator = selectionIndicator;
         this._infoBox = infoBox;
         this._dataSourceCollection = dataSourceCollection;
+        this._destroyDataSourceCollection = destroyDataSourceCollection;
         this._dataSourceDisplay = dataSourceDisplay;
         this._clockViewModel = clockViewModel;
         this._toolbar = toolbar;
@@ -792,11 +817,11 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
      * A mixin may add additional properties, functions, or other behavior
      * to the provided viewer instance.
      *
-     * @param {Function} mixin The Viewer mixin to add to this instance.
+     * @param {Viewer~ViewerMixin} mixin The Viewer mixin to add to this instance.
      * @param {Object} options The options object to be passed to the mixin function.
      *
      * @see viewerDragDropMixin
-     * @see viewerDynamicObjectMixin
+     * @see viewerEntityMixin
      */
     Viewer.prototype.extend = function(mixin, options) {
         //>>includeStart('debug', pragmas.debug);
@@ -911,7 +936,9 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
         this._dataSourceDisplay = this._dataSourceDisplay.destroy();
         this._cesiumWidget = this._cesiumWidget.destroy();
 
-        this._dataSourceCollection = this._dataSourceCollection.destroy();
+        if (this._destroyDataSourceCollection) {
+            this._dataSourceCollection = this._dataSourceCollection.destroy();
+        }
 
         return destroyObject(this);
     };
@@ -996,6 +1023,15 @@ Either specify options.terrainProvider instead or set options.baseLayerPicker to
         viewer._lastWidth = width;
         viewer._lastHeight = height;
     }
+
+    /**
+     * A function that augments a Viewer instance with additional functionality.
+     * @callback Viewer~ViewerMixin
+     * @param {Viewer} viewer The viewer instance.
+     * @param {Object} options Options object to be passed to the mixin function.
+     *
+     * @see Viewer#extend
+     */
 
     return Viewer;
 });
